Backend: + added game controller, service & repository + added RAWG database implementation
