+++ backend/src/main/java/de/jannisaziz/backend/game/GameController.java
@@ -1,7 +1,6 @@
-package de.jannisaziz.backend.controllers;
+package de.jannisaziz.backend.game;
 
-import de.jannisaziz.backend.models.Game;
-import de.jannisaziz.backend.services.GameService;
+import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -11,15 +10,12 @@ import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
+@AllArgsConstructor
 @RestController
 @RequestMapping("/api/games")
 public class GameController {
 
     private final GameService service;
-
-    public GameController(GameService service) {
-        this.service = service;
-    }
 
     @GetMapping("/search={name}")
     public List<Game> findGamesByName(@PathVariable String name) throws ResponseStatusException {
