+++ backend/src/main/java/de/jannisaziz/backend/game/GameRepository.java
@@ -1,6 +1,5 @@
-package de.jannisaziz.backend.repositories;
+package de.jannisaziz.backend.game;
 
-import de.jannisaziz.backend.models.Game;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
